some safety changes to bs412

+++ lib/bs412.h
@@ -12,4 +12,4 @@ float dbr_to_deviation(float dbr);
 float deviation_to_dbr(float deviation);
 
 void init_modulation_power_measure(MPXPowerMeasurement *mpx, int sample_rate);
-float measure_mpx(MPXPowerMeasurement *mpx, int deviation);
+float measure_mpx(MPXPowerMeasurement *mpx, float deviation);
